Implement default user for each book

+++ dao/BookDao.php
@@ -6,6 +6,7 @@
  * @author Roland Kluge
  */
 include_once('mapping/BookMapper.php');
+include_once('dao/UserDao.php');
 include_once('dao/Database.php');
 
 final class BookDao {
@@ -36,6 +37,7 @@ final class BookDao {
             $this->update($book);
         } else {
             $this->create($book);
+            $this->createDefaultUser($book);
         }
     }
 
@@ -57,10 +59,45 @@ final class BookDao {
         $statement->execute();
     }
 
-    public function drop($id) {
-        $sql = "DELETE FROM books WHERE id = :id";
+    public function addUserToBook(User $user, Book $book) {
+        $sql = "INSERT INTO users_to_books (user_id, book_id) VALUES (:user_id, :book_id)";
         $statement = Database::getDatabase()->prepare($sql);
-        $statement->execute(array(":id" => $id));
+        $statement->bindParam(":user_id", $user->getId());
+        $statement->bindParam(":book_id", $book->getId());
+        $statement->execute();
+    }
+
+    public function drop($id) {
+        $book = $this->get($id);
+        if ($book) {
+            $this->dropDefaultUser($book);
+
+            $sql = "DELETE FROM books WHERE id = :id";
+            $statement = Database::getDatabase()->prepare($sql);
+            $statement->execute(array(":id" => $id));
+        }
+    }
+
+    private function createDefaultUser(Book $book) {
+        $user = new User();
+        $user->setName("Gemeinschaft");
+        $user->setIsReal(false);
+
+        $userDao = new UserDao();
+        $userDao->save($user);
+
+        $this->addUserToBook($user, $book);
+    }
+    
+    private function dropDefaultUser(Book $book) {
+        $sql = "SELECT * FROM users_to_books JOIN users ON users_to_books.user_id = users.id WHERE book_id = :book_id AND NOT isReal";
+        $statement = Database::getDatabase()->prepare($sql);
+        $statement->bindParam(":book_id", $book->getId());
+        $statement->execute();
+        $row = $statement->fetch(PDO::FETCH_ASSOC);
+        
+        $userDao = new UserDao();
+        $userDao->drop($row['user_id']);
     }
 
 }
